Added ping_count option. Switched to min_rtt.
Switched to min_rtt because MacOS is doing something dumb. The first
ping is a very high value and then additional pings are realistic
values. This is skewing the results posted to the chart. Now the default
is to send two pings and report the min value instead of average.